--- Assets/App/Scripts/Main/Const.cs
@@ -1,63 +1,0 @@
-﻿//
-// Const.cs
-// ProductName Ling
-//
-// Created by toshiki sakamoto on 2019.05.01
-//
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using UnityEngine;
-using UnityEngine.UI;
-
-
-namespace Ling.Main
-{
-	/// <summary>
-	/// 
-	/// </summary>
-    public class Const
-    {
-        #region 定数, class, enum
-
-        public enum State
-        {
-            Start,
-            Adv, 
-        }
-
-        #endregion
-
-
-        #region public, protected 変数
-
-        #endregion
-
-
-        #region private 変数
-
-        #endregion
-
-
-        #region プロパティ
-
-        #endregion
-
-
-		#region コンストラクタ, デストラクタ
-
-		#endregion
-
-
-        #region public, protected 関数
-
-        #endregion
-
-
-        #region private 関数
-
-        #endregion
-    }
-}
