+++ Assets/App/Scripts/Map/Builder/Split.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc9cfea54da5f0f4ca01aa1c37e33fff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
