+++ Assets/App/Scripts/Map/Builder/Split/Normal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8ea4f38e12d6c4048b6f01e13856300a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
