+++ Assets/App/Scripts/Scenes/Main/Common.cs
@@ -1,0 +1,57 @@
+﻿//
+// Common.cs
+// ProductName Ling
+//
+// Created by toshiki sakamoto on 2019.05.01
+//
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using UnityEngine;
+using UnityEngine.UI;
+
+
+namespace Ling.Main
+{
+	/// <summary>
+	/// 
+	/// </summary>
+    public class Common
+    {
+        #region 定数, class, enum
+
+        #endregion
+
+
+        #region public, protected 変数
+
+        #endregion
+
+
+        #region private 変数
+
+        #endregion
+
+
+        #region プロパティ
+
+        #endregion
+
+
+		#region コンストラクタ, デストラクタ
+
+		#endregion
+
+
+        #region public, protected 関数
+
+        #endregion
+
+
+        #region private 関数
+
+        #endregion
+    }
+}
