+++ Assets/App/Scripts/Scenes/Main/MapView.cs
