+++ Assets/App/Tests/EditMode/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3b9617362305f6a4bba639b3f29c58dd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
