+++ Assets/App/Tests/EditMode/Editor/Utility/ListExtensionsTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 37ad8e3e0e098b94b998396b8a5ebc03
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
