+++ Assets/App/Tests/PlayMode/PlayMode.asmdef
@@ -1,0 +1,6 @@
+﻿{
+    "name": "PlayMode",
+    "optionalUnityReferences": [
+        "TestAssemblies"
+    ]
+}
